+++ packages/joi/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dual-autoform/joi
 
+## 1.0.1
+
+### Patch Changes
+
+- Build package declarations into `dist` and point package exports at the published declaration files.
+
 ## 3.0.1
 
 ### Patch Changes